requirements.txt
matplotlib
Pillow
numpy
rasterio
fastapi
uvicorn[standard]
python-multipart